REFACTOR FONT var and register_font function to image.py
