+++ pdf/modify/__init__.py
@@ -1,23 +1,8 @@
-import os
-from PyBundle import bundle_dir, resource_path
 from reportlab.lib.pagesizes import letter
-from reportlab.pdfbase import pdfmetrics
-from reportlab.pdfbase.ttfonts import TTFont
+from pdf.modify.draw.image import DrawPIL
 
 
-def register_font(font='Vera.ttf'):
-    """Register fonts for report labs canvas."""
-    directory = os.path.join(bundle_dir(), 'font')
-    ttfFile = resource_path(os.path.join(directory, font))
-    if os.path.exists(ttfFile):
-        pdfmetrics.registerFont(TTFont("Vera", ttfFile))
-        return ttfFile
-    else:
-        print(ttfFile, 'can not be found')
-
-
-FONT = register_font()
 LETTER = letter[1], letter[0]
 
 
-__all__ = ['LETTER', 'FONT']
+__all__ = ['LETTER', 'DrawPIL']
