+++ pdf/modify/draw/image.py
@@ -1,8 +1,23 @@
 import os
 from tempfile import NamedTemporaryFile
 from PIL import Image, ImageEnhance, ImageDraw, ImageFont
+from reportlab.pdfbase import pdfmetrics
+from reportlab.pdfbase.ttfonts import TTFont
 from PyBundle import bundle_dir, resource_path
-from pdf.modify import FONT
+
+
+def register_font(font='Vera.ttf'):
+    """Register fonts for report labs canvas."""
+    directory = os.path.join(bundle_dir(), 'font')
+    ttfFile = resource_path(os.path.join(directory, font))
+    if os.path.exists(ttfFile):
+        pdfmetrics.registerFont(TTFont("Vera", ttfFile))
+        return ttfFile
+    else:
+        print(ttfFile, 'can not be found')
+
+
+FONT = register_font()
 
 
 def img_adjust(image, opacity=1.0, rotate=None, fit=0, tempdir=None, bw=False):
